requirements.txt
Flask
requests
Pillow
colour